commit issues

+++ controllers/commentController.js
@@ -7,6 +7,7 @@ const upVote = require("../model/upVote");
 module.exports.commentController = async function(req,res){
 
     const user_type = req.query.type;
+    console.log(user_type);
     if(user_type=="User"){
         var UserOrHotel = "User"
     }
@@ -53,10 +54,10 @@ module.exports.commentController = async function(req,res){
     if(picture){
 
         if(user_type == "User"){
-            picturePath = `/uploads/userProfile/pics/${picture}`;
+            picturePath = `${picture}`;
         }
         else{
-            picturePath = `/uploads/hotelProfile/pics/${picture}`;
+            picturePath = `${picture}`;
     
         }
     }
@@ -66,7 +67,11 @@ module.exports.commentController = async function(req,res){
     }
 
 
-    console.log(picturePath);
+    // console.log(picturePath);
+    await req.flash('message', [
+        { type: 'flash-success', text: 'Comment Added' },
+      ]);
+
     if(req.xhr){
         
         return res.status(200).json({
@@ -77,26 +82,32 @@ module.exports.commentController = async function(req,res){
                 commentID:commentData.id,
                 typeOfCommenter:user_type,
                 picturePath:picturePath,
-                commentUpVoteCount:commentData.upVotesCount
+                commentUpVoteCount:commentData.upVotesCount,
+                countOfComments:postData.comments.length,
+                message: await req.flash('message')
             }
             
         })
 
     }     
-    
     return res.redirect('back');
     
 }
 
 module.exports.deleteComment = async function(req,res){
 
-    console.log("hello",req.body.cID,req.body.pID);
+    // console.log("hello",req.body.cID,req.body.pID);
 
 
     if(req.xhr){
         console.log("xhr");
 
-        let commentData = await comment.findById(req.body.cID);
+        let commentData = await comment.findById(req.body.cID)
+        .populate({
+            path:'user',
+            select:'id'
+        });;
+
 
         if(!commentData){
 
@@ -106,25 +117,44 @@ module.exports.deleteComment = async function(req,res){
 
         }
 
-        let postData = await post.findById(req.body.pID);
+        if(req.user.id == commentData.user.id || req.user.isAdmin){
+
+            let postData = await post.findById(req.body.pID);
+        
+            let commentUpvotes = commentData.upVotes;
+        
+            commentUpvotes.map(async data => {
+         
+                await upVote.findByIdAndDelete(data.toString());
+        
+            });
+        
+            await postData.comments.pull(commentData.id);
+        
+            await postData.save();
+        
+            await commentData.remove();
     
-        let commentUpvotes = commentData.upVotes;
-    
-        commentUpvotes.map(async data => {
-     
-            await upVote.findByIdAndDelete(data.toString());
-    
-        });
-    
-        postData.comments.pull(commentData.id);
-    
-        await postData.save();
-    
-        await commentData.remove();
-    
-        return res.status(200).json({
-            deleted:true
-        })
+            await req.flash('message', [
+                { type: 'flash-warning', text: 'Comment Deleted !' },
+              ]);
+        
+            return res.status(200).json({
+                deleted:true,
+                countOfComments:postData.comments.length,
+                postId:postData.id,
+                message: await req.flash('message')
+            })
+        }
+        else{
+            await req.flash('message', [
+                { type: 'flash-warning', text: "Can't delete" },
+              ]);
+        
+            return res.status(200).json({
+                message: await req.flash('message')
+            })
+        }
 
     }
